requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-prometheus==2.3.1
drf-yasg==1.21.7
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
Django>=4.2.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.1
celery>=5.3.0
redis>=4.5.0
django-filter>=23.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
drf-spectacular>=0.26.0
django-redis>=5.4.0
django-cacheops>=7.0.1
djangorestframework-simplejwt>=5.3.0
django-ratelimit>=4.1.0
elasticsearch-dsl>=7.4.0
django-elasticsearch-dsl>=7.2.2
prometheus-client>=0.16.0
django-prometheus>=2.3.1
python-json-logger>=2.0.7
drf-yasg==1.21.7
django-prometheus==2.3.1
celery==5.3.6
redis==5.0.1